planet details show and hide on click
