+++ src/javascripts/components/planetCards/planetCards.scss
@@ -17,20 +17,3 @@
         width: 100%
     }
 }
-
-#planetDetails{
-    color: white;
-    display: none;
-
-    button{
-        background-color: rgba(255, 255, 255, 0.377);
-        border: none;
-        border-radius: 5px;
-        color: red;
-        display: inline-block;
-        font-size: 2em;
-        font-weight: bold;
-        margin-left: 90%;
-        width: 2em;
-    }
-}
